@@ -1,26 +1,21 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<RootNamespace>Cronus.Monitor.Api</RootNamespace>
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Cronus" Version="7.0.0-preview.34" />
-		<PackageReference Include="Cronus.AspNetCore" Version="7.0.0-preview.3" />
-		<PackageReference Include="Cronus.Serialization.NewtonsoftJson" Version="7.0.0-preview.2" />
-		<PackageReference Include="Cronus.Transport.RabbitMQ" Version="7.0.0-preview.26" />
-		<PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.20.0" />
-		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="6.0.1" />
-		<PackageReference Include="Microsoft.CodeAnalysis.FxCopAnalyzers" Version="3.3.2" />
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.0" />
+		<PackageReference Include="Cronus" Version="10.10.0" />
+		<PackageReference Include="Cronus.AspNetCore" Version="10.2.1" />
+		<PackageReference Include="Cronus.Serialization.NewtonsoftJson" Version="10.4.0" />
+		<PackageReference Include="Cronus.Transport.RabbitMQ" Version="10.6.2" />
+		<PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.22.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="8.0.3" />
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
 	</ItemGroup>
 
 	<ItemGroup>
 		<ProjectReference Include="..\Log\Log.csproj" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<Folder Include="Infrastracture\" />
 	</ItemGroup>
 </Project>
