@@ -1,8 +1,7 @@
-﻿namespace Cronus.Monitor
+﻿namespace Cronus.Monitor;
+
+public static class App
 {
-    public static class App
-    {
-        public const string Name = "Cronus.Monitor";
-        public const string LogName = "cronus.monitor";
-    }
+    public const string Name = "Cronus.Monitor";
+    public const string LogName = "cronus.monitor";
 }
