@@ -2,21 +2,20 @@
 using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace Cronus.Monitor.Validations
+namespace Cronus.Monitor.Validations;
+
+[AttributeUsage(AttributeTargets.Property | AttributeTargets.Field | AttributeTargets.Parameter, AllowMultiple = false)]
+public class IsUrnAttribute : ValidationAttribute
 {
-    [AttributeUsage(AttributeTargets.Property | AttributeTargets.Field | AttributeTargets.Parameter, AllowMultiple = false)]
-    public class IsUrnAttribute : ValidationAttribute
+    public override bool IsValid(object value)
     {
-        public override bool IsValid(object value)
-        {
-            if (value is null) return true;
+        if (value is null) return true;
 
-            return Urn.IsUrn(value as string);
-        }
+        return Urn.IsUrn(value as string);
+    }
 
-        public override string FormatErrorMessage(string name)
-        {
-            return $"The field {name} is not a valid Urn. Urn fields must comply to the pattern: {UrnRegex.Pattern}";
-        }
+    public override string FormatErrorMessage(string name)
+    {
+        return $"The field {name} is not a valid Urn. Urn fields must comply to the pattern: {UrnRegex.Pattern}";
     }
 }
