@@ -1,47 +1,37 @@
 ﻿using Elders.Cronus;
-using Elders.Cronus.Multitenancy;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Cronus.Monitor.Api
+namespace Cronus.Monitor.Api;
+
+public class Worker : BackgroundService
 {
-    public class Worker : BackgroundService
+    private readonly ILogger<Worker> log;
+    private readonly ICronusHost cronusHost;
+
+    public Worker(ILogger<Worker> log, ICronusHost cronusHost)
     {
-        private readonly ILogger<Worker> log;
-        private readonly ICronusHost cronusHost;
+        this.log = log;
+        this.cronusHost = cronusHost;
+    }
 
-        public Worker(ILogger<Worker> log, ICronusHost cronusHost)
-        {
-            this.log = log;
-            this.cronusHost = cronusHost;
-        }
+    protected override async Task ExecuteAsync(CancellationToken stoppingToken)
+    {
+        log.LogInformation("Starting service...");
 
-        protected override async Task ExecuteAsync(CancellationToken stoppingToken)
-        {
-            log.LogInformation("Starting service...");
+        await cronusHost.StartAsync().ConfigureAwait(false);
 
-            cronusHost.Start();
+        log.LogInformation("Service started!");
+    }
 
-            log.LogInformation("Service started!");
-        }
+    public override async Task StopAsync(CancellationToken cancellationToken)
+    {
+        log.LogInformation("Stopping service...");
 
-        public override Task StopAsync(CancellationToken cancellationToken)
-        {
-            log.LogInformation("Stopping service...");
+        await cronusHost.StopAsync().ConfigureAwait(false); ;
 
-            cronusHost.Stop();
-
-            log.LogInformation("Service stopped");
-            return Task.CompletedTask;
-        }
-
-        public override void Dispose()
-        {
-            StopAsync(CancellationToken.None).GetAwaiter().GetResult();
-        }
+        log.LogInformation("Service stopped");
     }
 }
